# tianyi-forum-frontend/docs/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "天翼论坛 - 管理端 API",
    "version": "1.0.0",
    "description": "天翼论坛管理后台接口文档。所有业务响应统一返回 {code, msg, data} 的 Result 包装，业务数据放在 data 字段。",
    "contact": { "name": "天翼论坛开发团队" }
  },
  "servers": [
    { "url": "http://localhost:8080/api", "description": "本地开发环境" },
    { "url": "https://api.tianyi-forum.com/api", "description": "生产环境" }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "summary": "获取用户列表（分页）",
        "tags": ["用户管理"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "minimum": 1, "example": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "example": 10 }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": { "type": "string", "example": "user_10" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["active", "banned"], "example": "active" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Result" },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/User" }
                            },
                            "total": { "type": "integer", "example": 137 }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{userId}/status": {
      "put": {
        "summary": "更新用户状态（封禁/解封）",
        "tags": ["用户管理"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 10 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status", "reason"],
                "properties": {
                  "status": { "type": "string", "enum": ["active", "banned"], "example": "banned" },
                  "reason": { "type": "string", "example": "发布违规内容" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Result" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/SuccessResponse" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 1 },
          "msg": { "type": "string", "example": "success" },
          "data": { "type": "object" }
        },
        "description": "统一返回包装，业务数据放在 data 字段"
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 0 },
          "message": { "type": "string", "example": "success" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "username": { "type": "string", "example": "user_10" },
          "email": { "type": "string", "example": "user10@example.com" },
          "role": { "type": "string", "enum": ["user", "moderator", "admin"], "example": "user" },
          "status": { "type": "string", "enum": ["active", "banned"], "example": "active" },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15 10:30:00"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-28 14:20:00"
          }
        }
      },
...
    },
...
  }
}